refactored pmp to its own module
